// frontend/person-blog-app/packages/blog/src/shared/api/generated/models/userPostInfoModel.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Gateway.API
 * OpenAPI spec version: 1.0
 */
import type { PostVisibility } from './postVisibility';
import type { TextInfoDto } from './textInfoDto';
import type { VideoInfoDto } from './videoInfoDto';

export interface UserPostInfoModel {
  id?: string;
  blogId?: string;
  dislikeCount?: number;
  likeCount?: number;
  viewCount?: number;
  /** @nullable */
  paymentSubscriptionId?: string | null;
  visibility?: PostVisibility;
  /** @nullable */
  title?: string | null;
  textInfo?: TextInfoDto;
  videoInfo?: VideoInfoDto;
  createdAt?: string;
}